Bosh: Grab blobstore definition from meta stanza
We need to be able to override the bosh blobstore config with the S3
settings, but spruce does not easily allow to override a full hashmap
(see https://github.com/geofffranks/spruce#map-replacement)

To workaround the issue, we grab the blob store config from
`meta.blobstore`. This way all the entry can be easily overwritten.