Define in file tenant CIDRs with access to API+web
Terraform will read from a file defined by the variable
${var.tenant_cidrs_file}, one CIDR per line and allowing comments.

Prod will be the environment access for the actual tenants, so
for dev and staging it will load the file
terraform/tenant-cidrs-default.txt, and prod will load
terraform/tenant-cidrs-prod.txt as defined in terraform/prod.tfvars.

We move the office CIDRs to these files too, but keep the variable
office_cirds for more specific and operator related security groups.

The variable ${var.web_access_cidrs} is now empty, but will be updated
in prod to allow access from the world.