A ruby tool to synchronise admin users with CF
This tool takes the list of users from config/admin_users.yml and
synchronises them with the currently configured CF platform.

- Any missing users are added to the UAA
- New users are added as members of the groups defined in
  scripts/lib/uaa_sync_admin_users.rb
- Any users in the admin groups but not in the admin_users.yml file will
  be removed if they are older than two hours - this is to allow the
  acceptance/smoke tests to create users during the test run
- New users will recieve an email containing their admin login details
  encrypted to their GPG key (specified in admin_users.yml)

admin_users.yml is a YAML file with a list of hashes containing:
- email
- gpg_key: to encrypt the generated credentials
- username: optional, will default to email

We use webmock to test the class mocking UAA. All the interactions can be
recorded using `uaac` with the `--trace` option.

We note that while the majority of this code has tests we've explicitly
not tested the email sending code, as it is both extremely tricky and
basically testing that SES actually works. We should assume that SES
works.