Add ruby class to sync admin users users
Implement a class to interact with UAA and sync the Admin users with
a given list of users.

The consider an Admin user a member of the groups:
"cloud_controller.admin", "uaa.admin", "scim.read", "scim.write",
"doppler.firehose".

The class update method must:
 - Add any missing user to UAA.
 - Add the users as members of the given groups
 - Remove any user that is in the groups but not in the list.
   If the user is newer than 2h we do not delete, as it might be a
   tempory test user used in the acceptance/smoke tests.

We use webmock to test the class mocking UAA. All the interactions can be
recorded using `uaac` with the `--trace` option.