Add main sync-admin-users script
sync-admin-users.rb will connect to the given CF api urls, and interact
with UAA to create a list of admin users defined in the configuration
file.

The configuration file is a YAML file with a list of hashes containing:
 - email
 - gpg_key: to encrypt the generated credentials
 - username: optional, will default to email

All the created users will be added to the Admin groups defined in
scripts/lib/uaa_sync_admin_users.rb

The script will encrypt credentials and send a email to each one of
the newly created users.