Use real plan in tests, and remove temporary plan.
No what we have a real service plan, there's no need for the
temporary-test-plan that the tests were using.

It was thought that there would be issues using the new plan in the
tests because it's marked as a non-free plan. This turns out not to be
an issue because the acceptance test helpers already setup a quota that
allows access to non-free plans[1]

[1]https://github.com/cloudfoundry-incubator/cf-test-helpers/blob/41e1e26/helpers/context.go#L96