Plug admin user sync into the pipeline
User sync will only run on staging & prod envs - we disable it in dev
because we don't want to be emailing everybody whenever a new
environment is created.

We don't just turn off emailing as creating users when you don't tell
people the credentials is a bit pointless.

This uses a ruby 2.2 container because we need ruby, gpg, and build
tools as eventmachine needs to build a native extension, which means we
cannot use the slim container as it does not have the build tools
available.