List of admin users in admin_users.yml
This list was generated by running `cat .gpg-id | xargs -t -n1 gpg --list-keys`
in the root of the paas-cf repository and manually constructing the yaml
from the returned data.

We are not using the .gpg-id file for this as we need email address, and
we are not replacing the .gpg-id file with this as the list of users
with admin accounts is not likely to be the same list of users that can
merge & sign branches.