Add os-conf release and script_runner config
In order to address the aws-nat-gateway MTU issue until AWS deploy
their own fix, we are adding a script_runner job to our deployment
which will allow us to run a custom script on all VMs to force the
MTU value to 1500 bytes.

This commit adds the os-conf release into our deployment manifests
and configuration for the script_runner job to fix the MTU

This uses the new `addons` feature from BOSH 255.4 or later in order
to run a pre-execution script. There is no actual job for monit to run,
but we are still able to run a pre-execute script.

See: https://github.com/alphagov/paas-os-conf-release for more details.