Remove router static_ips
We are not sure why router had allocated static_ips (this likely comes
from times when router lived on the same subnet as core CF components).
We are removing the static ips as we can not find anything that
references these IPs any more. This also means we don't have to specify
migrated_from to preserve the IPs.