Log CF_TRACE output for tests in Concourse
During story #121039143 a need to look at CF_TRACE output from
tests was identified. It was discovered that setting the CF_TRACE
environment variable is not sufficient for the output to propagate
through to the Concourse UI. The documentation for cf-smoke-tests[1]
and cf-acceptance-tests[2] states it should be set via the
`artifacts_directory` configuration variable, which determines the
directory in which the trace output log files will be stored.

This work was no longer required before completion, but may be a useful
reference if a future story is raised to add CF_TRACE output from tests
to the pipeline.

[1] https://github.com/cloudfoundry/cf-smoke-tests#capturing-cf-cli-output
[2] https://github.com/cloudfoundry/cf-acceptance-tests#capturing-test-output