Disable acceptance tests in prod/staging
We want to disable the execution of acceptance tests in staging and
prod, and allow disable it in any other environment providing a flag.

The reason is that running the tests in prod/staging do not provide so
much value, and some of the acceptance tests are not safe to run in
production, as they occasionally fail in an unsafe way.

We add a new environment variable ENABLE_CF_ACCEPTANCE_TESTS which is
by default `true`, but `false` for prod and staging. The operator
can optionally enable this variable if required.

This will only stop the tests from running, but we will continue
configuring the jobs in concourse, and even execute all the preparation
steps: render configuration, prepare container, etc.

Importantly, in this commit we will keep creating and deleting
the temporary user. This will be disabled in a different commit.

We decided so because:
 * we want to minimise the changes required
 * we do not want to include any complex templating in the pipeline
   to conditionally include tasks.