Do not create tmp user in disabled acceptance test
When we not enable the acceptance tests, we don't want to create and
destroy and unnecesary temporary admin user. But we want to keep
using a share task across jobs for that operation.

In order to allow that, we add a new parameter for the create_admin and
delete_admin tasks: ENABLE_ADMIN_USER_CREATION. If set to "false" it
will skip the user creation. Dummy credential files will created
in the task output anyway.

With this semantics, we can use te ENABLE_CF_ACCEPTANCE_TESTS variable
to enable/disable the creation of the temporary user.