For further discussion
When rotating the grafana password we need to know what the existing one
is so we can login as admin to use the admin apis.

This is one approach to keeping the old password, so we can rotate when
secrets.grafana_admin_password_old exists and differs from
secrets.grafana_admin_password in a later job.

A better approach is probably in grafana space.  Maybe direct database
manipulation?