Stream the gorouter access logs to syslog
We want to get all the logs from the router in our logsearch platform.

Currently gorouter is configured to send the logs to syslog, but it is
not configured to stream the access logs.

By enabling this option, all the access_log  will be sent to syslog,
and from there to loggregator.