Configure metron_agent as an addon
We want to install the metron_agent in all the VMs, so that the all
the logs from syslog in all the are forwarded to the loggregrator
stack.

For that, we add the metron agent to the list of addon jobs.

But we cannot refer to the properties of loggregator to define
the dropsonde_incoming_port, so we extract this value to shared file:
manifests/shared/deployments/metron.yml