Restructure datadog terraform config
We are about to start writing a bunch of datadog checks.

Restructure beforehand to establish a convention that
each check goes into its own file.

Make a home for variables, and extract variable for
the deploy env into a `null_data_source` to establish
a convention for working around terraform's
inability to interpolate in variable declarations.
See https://github.com/hashicorp/terraform/issues/4084