Test push of python apps without Procfile
Test for the bug described in [1]: Staging fails for Python apps without
Procfile even though a command has been specified via manifest or `-c`
flag.

The bug was introduced in CF v245 with CAPI release 1.7.0, but fixed
using CAPI release 1.8.0[2]

The test will wait double when starting the python app, because we
observed that it often takes more than 30s.

[1] https://www.pivotaltracker.com/n/projects/966314/stories/132190561
[2] https://github.com/cloudfoundry/capi-release/releases/tag/1.8.0