Upload test artifacts to S3
We want to store the CF_TRACE output from tests run in the pipeline,
as they help with debugging test failures. We do not want them output
in the pipeline, as we cannot guarantee they won't contain secrets,
so we upload them to an S3 bucket.

The paths used to store the files are based from the way the upstream
cf-release does it. For the acceptance tests the path is hard coded[1].
For the smoke-tests they do not support configuring this value[2][3],
so a path consistent with that used by the acceptance tests was chosen.

[1] https://github.com/cloudfoundry/cf-release/blob/v245/jobs/acceptance-tests/templates/config.json.erb#L24
[2] https://github.com/cloudfoundry/cf-release/blob/v245/jobs/smoke-tests/templates/config.json.erb#L20-L37
[3] https://github.com/cloudfoundry/cf-release/issues/1089