Temporarily add redis port to broker App SG
This is needed to allow apps to connect to elasticache redis instances
deployed by the broker.

Note, while doing this, we ran into a YAML issue. the go-yaml library
(used by spruce), wrote the port value out without quotes. The ruby YAML
parser parsed this as a number (ignoring the comma), which then broke
cloudcontroller because it was expecting a string. Adding a space in
there causes spruce to insert the quotes, which avoids this problem