Add monitors for route-emitter lock
The metric cf.route-emitter.LockHeld.v1-locks-route_emitter_lock
measures whether a route-emitter holds the route-emitter lock: 1 means the lock
is held, and 0 means the lock was lost. Emitted every 30 seconds by the
active route-emitter.

As there can be only one route emitter, this metric can be only 1 for
each environment. Checking this value would allow us detect if there are
two route emitters thinking that are the master, as it happened
before[1].

The monitor must be exactly one, so we want to check upper and lower
thresholds. But Datadog does not allow that kind of check, and 2
different monitors must be created. We use `count=2` to create bosh and
we use `element()` and `format()` to build the right queries and
thresholds.

[1] https://github.com/cloudfoundry/diego-release/issues/201