Monitor total routers in route_emitter vs gorouter
This is a little bit complex monitor that checks that the gorouter is in
sync with the state in the diego route-emitter and the running LRPs in
Diego.

The route_emitter and gorouter total routes should be the same, except for
the instances that are desired but not running.

To monitor that, we compute the difference between the router-emitter
total routes minus the gorouter total routes, and subtract the
difference between desired LRP and running LRPs in BBS. The result must
be 0.

We want to check upper and lower thresholds, but Datadog does not allow
that. Because that we create two different monitors with `count=2` and
we use `element()` and `format()` to build the right queries.

Note: During my tests, in order to make this monitor work, I had to define
the monitor as "all values in last 15m" (aka `min(last_15m)`). Using a
smaller window would result in a "No Data" monitor. I opened a ticket to
learn why, but later the monitor was working with other window sizes.