Revert "Monitor total routers in route_emitter vs gorouter"
This reverts commit 95699bacb9cb85333e67ef40dda86b1cc598cca3.

The proposed monitor does not work as expected.

The number of DesiredLRPs - RunningLRPs do not represent the number of routes
ithout backend, as I initially assumed. Because that, pushing an app
that does not start without routes or with 2 or more routes would make
this monitor to fail.

In consequence we discard this monitor. We leave it as a commit for
reference and as a example.