Monitor if the route_emitter is refreshing routes
route_emitter connects to BBS and registers for events about changes
in the number of LRPs. When connected to the BBS it updates the
cf.route_emitter.RoutesRegistered and
cf.route_emitter.RoutesUnregistered every 30s.

If it is not connected, these metrics will not get updated.

This monitor will check that the metrics have been updated in the last
5 minutes. If not, it means that the route_emitter is not connected to
BBS.