Monitor if the route_emitter is syncing routes
route_emitter connects to BBS regulary to fully sync all the routes from
the BBS. When it does, it updates the metric:
cf.route_emitter.RouteEmitterSyncDuration.

If it is not connected, that metric will not get updated.

This monitor will check that the metric have been updated in the last
5 minutes. If not, it means that the route_emitter is not syncing with
the BBS.