Adjust conti. smoke test monitor no_data handling
Move no_data notification from monitor which only has failed jobs to monitor
which has all continuous smoke test runs. This is because 1st monitor gets
no_data when there are no failures - which is actually a good state we don't
want to be alerted about.

Adjust thresholds for no_data and slow jobs in slow conti monitor:
We now have m1.xlarge VM with 4 CPUs and 16G ram and moreover, concourse is
not getting stuck that often. This means that we get consistently fast runs
of smoke tests. We can thus lower thresholds to 12m warn and 20m alert. We
send pagerduty notification if there is no_data for more than 20 mins, as we
want to be alerted when our smoke-tests monitoring doesn't work any more.