Mock DataDog api
This function starts a http listener that returns DataDog specific json.
It will be used in update_monitor_options.rb test package to mock DD
api. It needs to go to BeforeSuite as there is no way to stop listener in test
context so next BeforeEach gets collision with already running goroutine.