Configure OAuth provider for UAA
We've decided to use MFA on our UAA accounts. The simplest solution is
to use one of the existing OAuth providers, such as Google.

UAA will act as a client and relying party on Google, so we need to extract
the client credentials from the password store and upload them to S3. This
will follow the same pattern as the uploading of Datadog credentials.