Use OpenID Connect authentication via Google
Pivotal story #136027405

We want to change from authenticating with UAA to authenticating using
OpenID Connect. The script will now delete users if their configured
origin is different to that held in UAA, and create them again with the
new origin. We need to delete the users because in UAA you can have
two users with the same username, as they will have different origins.

The newly created users should not have a password, as they should
only be allowed to authenticate via Google.