[PATCH] - Workaround for login issue[1]
This patch is an example of how to workaround the issue. It does
not include the updating of manifest unit tests.

Also, the CF CLI login prompt still seems to present you a passcode
URL that includes the old subdomain, I didn't bother tracking down
where it pulls that configuration from.

[1] https://github.com/cloudfoundry/uaa/issues/562