Only deploy datadog-consul on Consul VMs
The purpose of the datadog-consul job/template is to install a configuration
file on the VM, which the datadog agent uses to perform an HTTP check.
This check fails if the VM is not currently the Consul leader. Therefore,
it does not seem useful to for VMs that are not part of the Consul cluster
to perform this check, so we remove them from those VMs.

An issue has been encountered with the Datadog monitor that uses
these checks[1]. It has been failing intermittently and it is unclear
why. This change should not have a negative impact on the monitor (which
only fails if 100% of hosts running the check fail to report they are the
leader), but should reduce the noise and confounding information when
debugging such problems.

An alternative solution was proposed, which included configuring the job
so the check can be turned off (the configuration file would end up blank).
However, this means updating the paas-datadog-for-cloudfoundry-boshrelease
repository, then updating the properties in paas-cf to turn it off,
therefore complicating the process. Also, it would mean having a blank
configuration file on the host, and I don't know how the Datadog agent
would handle that.

[1] https://www.pivotaltracker.com/n/projects/1275640/stories/141948375