Monitor CPU credits for t2 instances
AWS t2 instance types have burstable CPU performance. They accumulate
"credits" when not doing anything and then can consume the credits when they
want to do CPU intensive work. When the credits run out, the instance will
not be able to do CPU intensive work and will slow down. We want to know
when this happens because it is normally correlates to another problem/alert
and may indicate that we need to change the instance type in future.

Writing this alert turned out to be much less trivial than I'd hoped. I no
longer feel so strongly about the usefulness of this, but it's probably
better than nothing and needing to lookup the information ourselves.

We have two metrics from the DataDog EC2 integration:

- `cpucredit_balance`: how many credits are available
- `cpucredit_usage`: how quickly an instance is consuming the balance

Each t2 instance type has slightly different characteristics in how many
credits they start of with, how they are used, how quickly they are
replenished, and how many can be kept in balance.

Ideally we would alert on the likelihood of the balance being exhausted in
the near future, but DataDog isn't able to do calculations based on two
metrics and we don't have (programmatic) information about the different
instance types.

So what we've settled on is an informational alert that will tell us when an
instance is very close to or has already exhausted it's credits. This will
allow us to correlate the information with other alerts that may be related.
The initial balance is 30 credits for `t2.{nano,micro,small}` instances and
60 credits for `t2.{medium,large}` instances, so I've set the warning
threshold to be below the minimum. The value and window should prevent
false-positives from newly created instances.

The edit page for the monitor warns about the following:

> Note: The selected metric provides data at low frequencies, we strongly recommend alerting on a 1 hour timeframe.

I haven't used 1h because it seems very large and will take a long time for
the monitor to either trigger or recover. The data points seem to be 1~3mins
apart, which doesn't seem terrible, so I've compromised on a window of 30m.

I've used `avg():` instead of `min():` so that it doesn't take an incredibly
long time for the monitor to recover when credits are replenished. I don't
think the second `:avg:` has any effect because the `by` clause separates
each instance.

We have decided not to add a separate monitor for RDS at this stage because
we don't currently have any t2 instances for of our Cloud Foundry
installation. If we do add it the future then it's worth noting that the
metrics appear to be tagged only with `aws_account` and not `deploy_env`
like the EC2 metrics are.