TMP show why I think healthcheck logic needs tewaking
Previously there was one switch statement which tried to set both the default
service and set the necessary values for turning SSL on/off in both drivers.

But as well as the intended side effect of default service to postgres when it
wasn't passed as a url parameter, it also had the effect of silently accepting
other values:

$ curl https://healthcheck-graham.cloudapps.digital/db?service=neo4j
{
  "success": true
}

Behaviour after this commit:

$ curl https://healthcheck-graham.cloudapps.digital/db?service=neo4j
unknown service: neo4j
$

Not only that but I find it easier to understand what the switch statement is
doing now.