Expect to route through an ALB on a private subnet
The final entry in the `X-Forwarded-For` header in a request received
by an app should be the IP of the ALB that it passed through. The ALB
will vary, as will its IP address, but we expect it to be on a private
subnet. This test is updated with this expectation.