Use same update-serial than in cf-deployment
Use the same settings for the update policy for each instance_group
as the ones defined in cf-deployment. We understand that is the
blessed update order and serialisation.

In cf-deployment no instance has `max_in_flight` set. We think it is
save to remove it because in all these cases we deploy across
multiple AZs as many instances as AZs. BOSH would update one
instance per AZ and max_in_flight would have no effect.
Additionally, in some cases (e.g. API) the `max_in_flight` is
incorrectly under `properties: ` instead of `update: `.