Use bosh with cf-vars-store instead of cf-certs
We want to start using the `bosh interpolate` certificate generation,
instead of the generated cf-certs.tar.gz that used certstrap.

In this commit we start consuming the cf-vars-store.yml file that
stores the variables/certificates generated by bosh. We start
consuming the version migrated from the old cf-certs from the job
`migrate-cf-certs`.

As `bosh interpolate` might update the vars-store file we push it
back once the task is finished.

We pass the ipsec-CA as a variable for the time being, as this
certificate is not part of cf-deployment and we consider that
it is better to test it independently and ensure proper rotation.

We move the `--var-errs` to the pipeline, to prevent the script
from failing if `--vars-store` is not passed.