Support CA certificate rotation in the manifest
In order to support rotation of the CA certificates, we add
multiple certificates called `<name>_old`, that will hold the old
certificate after rotation.

In every service that consume these CA certificates, we concatenate
the current and old one.

During rotation, we will use a script that  would rename the current
certificate to the  `<name>_old`, or set `<name>_old.certificate`
to empty string.