Add the uaa_ca as a ca_cert to the go router
We have UAA behind the gorouter using SSL. So far this is working
OK because the bosh-CA is added as a root trusted certificate in
every VM, including the gorouter.

But if UAA cert would be using a different CA, we need
to have gorouter trusting it. To do so we populate the property
router.ca_certs.