Update certificate rotation to use vars store
Introduce a new script to modify the vars store where certs are now
generated and stored. The new tasks need to be run in the following
painstaking order:

1. Run `rotate-cf-certs-cas` to copy existing CAs into `_old` suffixes
1. Run `create-cloudfoundry` to generate new CAs and deploy them
   alongside the old CAs
1. Run `rotate-cf-certs-leafs` to delete existing non-CA certs
1. Run `create-cloudfoundry` to generate new non-CA certs against the
   new CAs and deploy them
1. Run `delete-old-cf-certs` to delete old CAs
1. Run `create-cloudfoundry` to remove the old CAs

It's hard to test Ruby CLI applications from rspec so I've focused on
testing the main function and parsing the fixtures from YAML so that
they are representative and can be copy/paste from the real manifests.