Use a 300-second timeout for the system domain ELB
We are moving the Cloud Controller API behind the Gorouter and behind
the system domain ELB. Deploying this the first time, we discovered
that the system domain ELB had a 19-second timeout. This caused
long-running CC API tasks (e.g., deploying app) to fail.

This commit removes the 19-second timeout and replaces that with the
300-second timeout already used by the apps domain ELB.

This change is experimental. We believe the 19-second timeout was
created when UAA was directly behind this ELB, as a measure to allow
zero-downtime redeploys of UAA. More recently UAA was put behind the
Gorouter, meaning that the ELB no longer needs to handle its downtime.

To prove this change we will verify availability when redeploying CC,
UAA and Gorouters.