Increase idle timeout for the system domain ELB
We are moving the Cloud Controller API behind the Gorouter and behind
the system domain ELB. Deploying this the first time, we discovered that
the system domain ELB had a 19-second timeout. This caused long-running
CC API tasks (e.g., deploying app) to fail.

We increase to the same value as the CloudController's current ELB.
This may need changing, given traffic now goes via GoRouter, but this
will be a good starting point.