Add deployment of prometheus to monitor CF
We add the deployment of prometheus, using the prometheus
bosh-release[1] with the default configuration for these
features enabled:

 - bosh exporter: To monitor the bosh metrics, jobs/vms, etc.
   it also enables autodiscovery[2]
 - cf exporter: To get the CF metrics and connect to firehose.

We deploy prometheus is a separated deployment in bosh, using
the `create-cloudfoundry` pipeline after `cf-deploy`.

We use bosh-interpolate password generation, storing the
`vars-store` file in S3.

We create the UAA clients for the CF exporter by using the opsfile
provided by prometheus bosh-release. The passwords are generated
by bosh-interpolate and we consume them by passing the CF `vars-store`
file as a `vars-file` when generating the prometheus manifest.

We have to add a `vm_extension` to add the required security group
to allow access to the bosh API from the prometheus VM.

All the changes of vm_types, network and vm_extensions is done
in a opsfile in this repo.

[1] https://github.com/bosh-prometheus/prometheus-boshrelease
[2] https://github.com/bosh-prometheus/prometheus-boshrelease#using-bosh-service-discovery