Add test for app pushes that take a long time
When initially trying to move Cloud Controller behind GoRouter[1] we
had test failures due to the short idle timeout on the ELBs. The tests
only happened to have caught the failure due to a `cf push` that took a
long time. This adds an explicit test for this, as we can't rely on existing
tests to always trigger this kind of failure.

[1] https://www.pivotaltracker.com/n/projects/1275640/stories/155856825