Update vm_types for different instance_groups
We have defined alternate vm_types in several instance groups,
to scale them up or due historical reasons.

We need to review some of these values, like the router vm_type,
which had its own vm_type to connect to a ELB. This is not required
anymore thx to the vm_extensions.