Setup custom session secret
We're sneaking in a quick fix for the session secret which previously
has been set to default.

This will invalidate cookie session for every user. It should do no
harm, as these expire after 12h anyway. If the session is invalid, the
user will be prompted to log in once again - which is both good and
desired.