wip: Hardcode log-cache/reverse_log_proxy link
Previously this defaulted to a `.bosh` hostname, which doesn't work
because we don't currently use BOSH DNS and we don't want to rush
enabling it while it's still "experimental".

I tried using `ip_addresses: true` (https://bosh.io/docs/dns/#links) but
it had no effect, possibly because the instance group is on a dynamic
network.

You have to specify `instances` when specifying `properties`, but it can
be an empty slice because the release uses `address` instead
(https://bosh.io/docs/links-manual/)

Consul agent needs to be run on the same instance group in order to
resolve the internal service hostname that RLP publishes.