Add Bosh DNS release as addon
This will install the Bosh DNS release on all VMs. It is to investigate
the use of Bosh DNS.

Configured as per the documentation[1]. We have turned off healthchecks
for now because we have no need for it during the spike. The release
seems to require API TLS configuration[2], but we are not sure why at
this stage.

[1] https://bosh.io/docs/dns/#enable
[2] https://github.com/cloudfoundry/bosh-dns-release/blob/master/jobs/bosh-dns/spec#L49-L52