Swap the old-new certificate order when rotating certificates.
Currently the CA cert rotation is broken. The root of the problem is that when
OpenSSL is configured to trust multiple CAs, and two of them have the same
subject name, OpenSSL will only verify certificates against the first one (see
[1] in OpenSSL code).

We updated BOSH cli to add Subject Key Identifiers to newly generated
certificates so OpenSSL is always able to choose the right one even if there
are multiple with the same subject name.

On the first rotation the old certificates still won't have SKIs, so we have to
make sure the old certificates are the first, so OpenSSL will still use them.

[1] https://github.com/openssl/openssl/blob/49f6cb968ff63793f6671d9026fb2a7034dad79a/crypto/x509/x509_lu.c#L613-L617