Stop using dora in custom-acceptance-tests
Dora is an app used in the upstream cf-acceptance-tests to test some
http behaviours. We were using it for a single test here, and had our
own app for all the other cases we were testing. Adding this case to our
app simplifies the tests.

This also removes this dependency on cf-acceptance-tests which is
non-obvious, and also means that we can remove some dependencies from
the pipeline.