Remove Consul and Consul agent
We intend to remove Consul as we no longer use it for locking or DNS.
This is based on the experimental ops file for removing Consul. We only
remove the things which add Consul, its agent, or configuration, which
would break if not removed. More Consul related configuration will be
removed in upcoming commits.